Factor out `CheckboxWithLabel` component
